+++ noodles-sam/src/reader/record/data.rs
@@ -3,11 +3,13 @@ pub(crate) mod field;
 use std::{error, fmt};
 
 use self::field::parse_field;
-use crate::record::Data;
+use crate::record::{data::field::Tag, Data};
 
 /// An error when raw SAM record data fail to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
+    /// A tag is duplicated.
+    DuplicateTag(Tag),
     /// A field is invalid.
     InvalidField(field::ParseError),
 }
@@ -15,6 +17,7 @@ pub enum ParseError {
 impl error::Error for ParseError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
+            Self::DuplicateTag(_) => None,
             Self::InvalidField(e) => Some(e),
         }
     }
@@ -23,23 +26,27 @@ impl error::Error for ParseError {
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            Self::DuplicateTag(tag) => write!(f, "duplicate tag: {tag}"),
             Self::InvalidField(e) => {
                 write!(f, "invalid field")?;
 
                 if let Some(tag) = e.tag() {
                     write!(f, ": {tag}")?;
                 }
+
+                Ok(())
             }
         }
-
-        Ok(())
     }
 }
 
 pub(crate) fn parse_data(mut src: &[u8], data: &mut Data) -> Result<(), ParseError> {
     while !src.is_empty() {
         let (tag, value) = parse_field(&mut src).map_err(ParseError::InvalidField)?;
-        data.insert(tag, value);
+
+        if let Some((t, _)) = data.insert(tag, value) {
+            return Err(ParseError::DuplicateTag(t));
+        }
     }
 
     Ok(())
@@ -70,6 +77,12 @@ mod tests {
         parse_data(b"NH:i:1\tCO:Z:ndls", &mut data)?;
         let expected = [nh, co].into_iter().collect();
         assert_eq!(data, expected);
+
+        data.clear();
+        assert_eq!(
+            parse_data(b"NH:i:1\tNH:i:1", &mut data),
+            Err(ParseError::DuplicateTag(tag::ALIGNMENT_HIT_COUNT))
+        );
 
         data.clear();
         assert!(matches!(